+++ src/main/java/tave/crezipsa/crezipsa/application/file/dto/response/DownloadUrlResponse.java
@@ -1,0 +1,5 @@
+package tave.crezipsa.crezipsa.application.file.dto.response;
+
+public record DownloadUrlResponse(String downloadUrl) {
+
+}
