+++ src/main/java/tave/crezipsa/crezipsa/application/file/dto/response/UploadUrlResponse.java
@@ -1,4 +1,4 @@
-package tave.crezipsa.crezipsa.presentation.file.response;
+package tave.crezipsa.crezipsa.application.file.dto.response;
 
 public record UploadUrlResponse(String fileName, String uploadUrl) {
 }
