--- src/main/java/tave/crezipsa/crezipsa/presentation/file/response/DownloadUrlResponse.java
@@ -1,5 +1,0 @@
-package tave.crezipsa.crezipsa.presentation.file.response;
-
-public record DownloadUrlResponse(String downloadUrl) {
-
-}
